Create non-existent instances is create=True
It's possible that the object related with the cache model that we are
trying to access doesn't exist yet.

If we had created our caching model with create=True, example:

    cache_relation(User.settings, create=True)

the instance for User.settings is going to be created in caste that it
was not found.

Added some spaces to be Flake8 compliant.