Cope with only some of the related models actually being loaded
The main thing this fixes is the case where the primary model is
not returned in the cached results. However it also changes the
handling in the case where one of the related models isn't found.
Previously the former would error and the latter would silently
assign `None` into the local relation cache field (and would thus
never be populated even if there should have been a value there).